+++ src/ApiService/ApiService/onefuzzlib/Nodes.cs
@@ -1,20 +1,17 @@
-﻿using ApiService.onefuzzlib.orm;
+﻿using Microsoft.OneFuzz.Service.OneFuzzLib.Orm;
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace ApiService
+namespace Microsoft.OneFuzz.Service;
+
+public partial record Node
 {
-    public partial record Node
-    {
-        public async static Task<Node?> GetByMachineId(IStorageProvider storageProvider, Guid machineId) {
-            var tableClient  = await storageProvider.GetTableClient("Node");
+    public async static Task<Node?> GetByMachineId(IStorageProvider storageProvider, Guid machineId) {
+        var tableClient  = await storageProvider.GetTableClient("Node");
 
-            var data = storageProvider.QueryAsync<Node>(filter: $"RowKey eq {machineId}");
+        var data = storageProvider.QueryAsync<Node>(filter: $"RowKey eq {machineId}");
 
-            return await data.FirstOrDefaultAsync();
-        }
+        return await data.FirstOrDefaultAsync();
     }
 }
