+++ src/ApiService/ApiService/onefuzzlib/orm/Orm.cs
@@ -4,15 +4,11 @@ using System.Reflection;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Text.Json;
-using ApiService;
-using System.Threading.Tasks;
-using System.Collections.Generic;
-using ApiService.onefuzzlib.orm;
 using System.Text.Json.Serialization;
 using System.Collections.Concurrent;
 using Azure;
 
-namespace Microsoft.OneFuzz.Service;
+namespace Microsoft.OneFuzz.Service.OneFuzzLib.Orm;
 
 public abstract record EntityBase
 {
